development --version 2.0 (added monitoring system and docker containerization)
