+++ .idea/encodings.xml
@@ -6,6 +6,8 @@
     <file url="file://$PROJECT_DIR$/api_server_gateway/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/discovery_service_server/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/inventory_service/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/notification-service/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/notification_service/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/order_service/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/order_service/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/product_service/src/main/java" charset="UTF-8" />
