+++ .idea/misc.xml
@@ -11,6 +11,7 @@
       <set>
         <option value="$PROJECT_DIR$/api_gateway/pom.xml" />
         <option value="$PROJECT_DIR$/api_server_gateway/pom.xml" />
+        <option value="$PROJECT_DIR$/notification-service/pom.xml" />
       </set>
     </option>
   </component>
