+++ api-gateway-server/Dockerfile
@@ -1,0 +1,5 @@
+FROM openjdk:17
+
+COPY target/*.jar app.jar
+
+ENTRYPOINT ["java","-jar","/app.jar"]
