+++ discovery_service_server/src/main/resources/application.yml
@@ -6,6 +6,10 @@ spring:
     name: discovery-service-server
 
 management:
+  endpoints:
+    web:
+      exposure:
+        include: prometheus
   tracing:
     sampling:
       probability: 1.0
