+++ notification_service/src/main/java/ca/tamrakar/notifcation_service/events/OrderPlacedEvent.java
@@ -1,0 +1,12 @@
+package ca.tamrakar.notifcation_service.events;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class OrderPlacedEvent {
+    private String orderNumber;
+}
