+++ notification_service/src/main/resources/application.yml
@@ -1,0 +1,24 @@
+server:
+  port: 9002
+
+hoster: localhost:7092
+
+spring:
+  application:
+    name: notification-service
+    zipkin:
+      base-url: http://localhost:9411
+
+management:
+  endpoints:
+    web:
+      exposure:
+        include: prometheus
+  tracing:
+    sampling:
+      probability: 1.0
+
+eureka:
+  client:
+    serviceUrl:
+      defaultZone: http://eureka:password@localhost:8761/eureka
