--- order_service/src/main/java/com/tamrakar/api_gateway_server/order_service/OrderServiceApplication.java
@@ -1,13 +1,0 @@
-package com.tamrakar.api_gateway_server.order_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OrderServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OrderServiceApplication.class, args);
-	}
-
-}
