--- order_service/src/main/java/com/tamrakar/api_gateway_server/order_service/config/WebClientConfig.java
@@ -1,16 +1,0 @@
-package com.tamrakar.api_gateway_server.order_service.config;
-
-import org.springframework.cloud.client.loadbalancer.LoadBalanced;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@Configuration
-public class WebClientConfig {
-
-    @Bean
-    @LoadBalanced
-    public WebClient.Builder webClient() {
-        return WebClient.builder();
-    }
-}
