--- order_service/src/main/java/com/tamrakar/api_gateway_server/order_service/model/InventoryResponse.java
@@ -1,15 +1,0 @@
-package com.tamrakar.api_gateway_server.order_service.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class InventoryResponse {
-    private String itemName;
-    private boolean isAval;
-}
