--- order_service/src/main/java/com/tamrakar/api_gateway_server/order_service/repository/OrderRepository.java
@@ -1,9 +1,0 @@
-package com.tamrakar.api_gateway_server.order_service.repository;
-
-import com.tamrakar.api_gateway_server.order_service.entity.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OrderRepository extends JpaRepository<Order, Long> {
-}
