+++ order_service/src/main/java/com/tamrakar/order_service/OrderServiceApplication.java
@@ -1,0 +1,13 @@
+package com.tamrakar.order_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OrderServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OrderServiceApplication.class, args);
+	}
+
+}
