+++ order_service/src/main/java/com/tamrakar/order_service/config/KafkaConfig.java
@@ -1,0 +1,33 @@
+package com.tamrakar.order_service.config;
+
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.tamrakar.order_service.events.OrderPlacedEvent;
+import org.apache.kafka.clients.producer.ProducerConfig;
+import org.apache.kafka.common.serialization.StringSerializer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.kafka.core.DefaultKafkaProducerFactory;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.kafka.core.ProducerFactory;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class KafkaConfig {
+    @Bean
+    public ProducerFactory<String, OrderPlacedEvent> producerFactory()
+    {
+        Map<String,Object> config = new HashMap<>();
+
+        config.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:29092");
+        config.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
+        config.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
+
+        return  new DefaultKafkaProducerFactory<>(config);
+    }
+
+    @Bean
+    public KafkaTemplate kafkaTemplate()
+    {
+        return  new KafkaTemplate(producerFactory());
+    }
+}
