+++ order_service/src/main/java/com/tamrakar/order_service/config/WebClientConfig.java
@@ -1,0 +1,16 @@
+package com.tamrakar.order_service.config;
+
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@Configuration
+public class WebClientConfig {
+
+    @Bean
+    @LoadBalanced
+    public WebClient.Builder webClient() {
+        return WebClient.builder();
+    }
+}
