+++ order_service/src/main/java/com/tamrakar/order_service/dto/OrderLineItemsDto.java
@@ -1,0 +1,17 @@
+package com.tamrakar.order_service.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class OrderLineItemsDto {
+    private Long orderLineItemsId;
+    private String orderLineItemsName;
+    private BigDecimal orderLineItemsPrice;
+    private int orderLineItemsQuantity;
+}
