+++ order_service/src/main/java/com/tamrakar/order_service/events/OrderPlacedEvent.java
@@ -1,0 +1,12 @@
+package com.tamrakar.order_service.events;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class OrderPlacedEvent {
+    private String orderNumber;
+}
