+++ order_service/src/main/java/com/tamrakar/order_service/model/InventoryResponse.java
@@ -1,0 +1,15 @@
+package com.tamrakar.order_service.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class InventoryResponse {
+    private String itemName;
+    private boolean isAval;
+}
