+++ order_service/src/main/java/com/tamrakar/order_service/model/OrderRequest.java
@@ -1,0 +1,15 @@
+package com.tamrakar.order_service.model;
+
+import com.tamrakar.order_service.dto.OrderLineItemsDto;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class OrderRequest {
+    private List<OrderLineItemsDto> orderLineItemsDto;
+}
