+++ order_service/src/main/java/com/tamrakar/order_service/repository/OrderRepository.java
@@ -1,0 +1,9 @@
+package com.tamrakar.order_service.repository;
+
+import com.tamrakar.order_service.entity.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderRepository extends JpaRepository<Order, Long> {
+}
