+++ order_service/src/main/java/com/tamrakar/order_service/service/OrderService.java
@@ -1,0 +1,9 @@
+package com.tamrakar.order_service.service;
+
+import com.tamrakar.order_service.model.OrderRequest;
+
+public interface OrderService {
+
+    String test();
+    boolean placeOrders(OrderRequest orderRequest);
+}
