+++ order_service/src/main/resources/application.yml
@@ -15,8 +15,24 @@ spring:
     properties:
       hibernate:
         dialect: org.hibernate.dialect.MySQL8Dialect
+#  kafka:
+#    bootstrap-servers: localhost:9092
+#    template:
+#      default-topic: notificationTopic
+#    producer:
+#      key-serializer: org.apache.kafka.common.serialization.StringSerializer
+#      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
+#      properties:
+#        spring:
+#          json:
+#            type:
+#              mapping: event:com.tamrakar.order_service.events.OrderPlacedEvent
 
 management:
+  endpoints:
+    web:
+      exposure:
+        include: prometheus
   tracing:
     sampling:
       probability: 1.0
