--- order_service/src/test/java/com/tamrakar/api_gateway_server/order_service/OrderServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.tamrakar.api_gateway_server.order_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OrderServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
