+++ order_service/src/test/java/com/tamrakar/order_service/OrderServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.tamrakar.order_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OrderServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
